Use AppConfig to determine events resolve behavior in view mode
Adds an AppConfig setting for max_linked_events_to_resolve. When
the max is exceeded (> no. of linked events) then the events and
linked records are not resolved in view mode and a simplified
display is presented (excludes titles). This keeps memory usage
under control as the no. of linked events grows.